+++ src/index.js
@@ -13,6 +13,7 @@ export {
 export { RowanAlert } from "./alert/alert.js";
 export { RowanAppLayout } from "./app-layout/app-layout.js";
 export { RowanBulkActionsBar } from "./bulk-actions-bar/bulk-actions-bar.js";
+export { RowanBulletChart } from "./bullet-chart/bullet-chart.js";
 export { applyFilters, RowanFilterBuilder } from "./filter-builder/filter-builder.js";
 export { RowanButton } from "./button/button.js";
 export { RowanCard } from "./card/card.js";