+++ types/index.d.ts
@@ -5,6 +5,7 @@ export { keys } from "./lib/keys.js";
 export { RowanAlert } from "./alert/alert.js";
 export { RowanAppLayout } from "./app-layout/app-layout.js";
 export { RowanBulkActionsBar } from "./bulk-actions-bar/bulk-actions-bar.js";
+export { RowanBulletChart } from "./bullet-chart/bullet-chart.js";
 export { RowanButton } from "./button/button.js";
 export { RowanCard } from "./card/card.js";
 export { RowanCheckbox } from "./checkbox/checkbox.js";